requirements.txt
flask
flask-sqlalchemy
flask-session
python-dotenv
flask_bcrypt
marshmallow
apispec
flask_apispec
flask_restful
flask-cors
flask-mysql
mysqlclient
redis
authlib
APScheduler
flask-APScheduler
requests